gameConfigクラスを別ファイルで定義した
