+++ Scripts/gameConfig.js
@@ -1,0 +1,5 @@
+/**
+ * ゲームの設定
+ */
+export class GameConfig {
+}
