+++ Scripts/gameConfig.ts
@@ -1,0 +1,11 @@
+/**
+ * ゲームの設定
+ */
+export class GameConfig {
+    public row: number;
+    public column: number;
+    public jsonPath: string;
+    public imageFolderPath: string;
+    public flippingWaitTimeMilliseconds: number;
+    public maxSelectableCard: number;
+}
